# Microoptimizations