- Replace two map lookups with one in the villager trade implementation
- Cache the hashCode of ItemIDs to avoid repeatedly calling into the somewhat slow Objects.hashCode
- Use actual equality comparisons instead of hashCode calculations for equals(), comparing on average 3 pointers that are next to each other in memory will be faster than computing hashCodes

Before (in an empty devenv, gains in a real environment should be higher):
```
Recipe map generated! Mapped 30 entities! It took 724ms
Villager trade information generation took 5 ms
```

After:
```
Recipe map generated! Mapped 30 entities! It took 573ms
Villager trade information generation took 4 ms
```